fix: fix clubs list button flicker
